+++ Classes/SalutationSwitcher.php
@@ -83,19 +83,13 @@ abstract class Tx_Oelib_SalutationSwitcher extends AbstractPlugin
      * 'greeting'.
      *
      * @param string $key the local language key for which to return the value, must not be empty
-     * @param bool $useHtmlSpecialChars @deprecated will be removed in oelib 3.0
      *
      * @return string the requested local language key, might be empty
      */
-    public function translate($key, $useHtmlSpecialChars = false)
+    public function translate($key)
     {
         if ($key === '') {
             throw new \InvalidArgumentException('$key must not be empty.', 1331489025);
-        }
-
-        if ($useHtmlSpecialChars) {
-            // will be removed in oelib 3.0.0
-            GeneralUtility::logDeprecatedFunction();
         }
 
         if (isset($this->translationCache[$key])) {